requirements.txt
RPLCD
requests
python-dotenv
smbus2
python-telegram-bot
flask
adafruit-circuitpython-dht
RPi.GPIO